Fix a couple bugs with map reading

+++ path_planner/src/common/map/GridWorldMap.cpp
@@ -96,6 +96,6 @@ const double* GridWorldMap::extremes() const {
     return m_Extremes;
 }
 
-int GridWorldMap::resolution() const {
+double GridWorldMap::resolution() const {
     return m_Resolution;
 }
