+++ path_planner/src/common/map/GridWorldMap.h
@@ -21,11 +21,11 @@ public:
 
     const double* extremes() const override;
 
-    int resolution() const override;
+    double resolution() const override;
 
 private:
     std::vector<std::vector<bool>> m_Blocked;
-    int m_Resolution;
+    double m_Resolution;
     double m_Extremes[4];
 };
